refactor(labtest): extract indicator table and modal components
- Add IndicatorTable component for grouped indicator display
- Add IndicatorEditModal component for add/edit functionality
- Reduce main file from 547 to ~320 lines
